Iss#44 - lot of refactoring in progress
