+++ src/js/models/meeting-minutes.js
@@ -1,0 +1,20 @@
+import Model from './model';
+
+export default class DatabaseEntryModel extends Model{
+	constructor(data) {
+		super();
+		this.obj = Object.assign(data);
+	}
+	// get obj(){
+	// 	return this.obj;
+	// }
+	// set obj(data){
+	// 	console.log("In the obj set method?");
+	// 	this.obj = Object.assign(data);
+	// }
+	
+	
+	
+	
+	
+}
