+++ src/js/routes/meeting-minutes.js
@@ -1,11 +1,11 @@
 import MeetingMinutesController from '../controllers/meeting-minutes';
 import AuthenticationService from '../services/authentication';
-import ModalController from '../controllers/showModal';
+import DatabaseModel from '../models/meeting-minutes';
 
 
 //Shows Member Account Information on the Members Page
 export default function meetingMinutes(ctx, next) {
-	new MeetingMinutesController(new AuthenticationService(), ModalController);
+	new MeetingMinutesController(new AuthenticationService(), DatabaseModel);
 	
 	next();
 }
